refactor(hive): Extract and compile generator
